# backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LoginResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string",
						"example": "usuario@email.com"
					},
					"password": {
						"type": "string",
						"example": "senhaSegura123"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cpf": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"ADMIN"
						]
					}
				},
				"required": [
					"id",
					"cpf",
					"phoneNumber",
					"name",
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDTO": {
				"properties": {
					"cpf": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"cpf",
					"phoneNumber",
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_UserResponseDTO_": {
				"properties": {
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/UserResponseDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"totalItems",
					"totalPages",
					"currentPage",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDTO": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SaleResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"valueTotal": {
						"type": "number",
						"format": "double",
						"example": 150.5
					},
					"description": {
						"type": "string",
						"example": "Venda de periféricos"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"example": 1
					}
				},
				"required": [
					"id",
					"valueTotal",
					"description",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaleItemRequest": {
				"properties": {
					"itemId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"quantity": {
						"type": "number",
						"format": "double",
						"example": 2
					}
				},
				"required": [
					"itemId",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaleCreateRequest": {
				"properties": {
					"description": {
						"type": "string",
						"example": "Venda de periféricos gamer"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/SaleItemRequest"
						},
						"type": "array",
						"example": [
							{
								"itemId": 1,
								"quantity": 2
							}
						]
					}
				},
				"required": [
					"description",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaleDetailResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"valueTotal": {
						"type": "number",
						"format": "double",
						"example": 150.5
					},
					"description": {
						"type": "string",
						"example": "Venda de periféricos"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"vendedor": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"name"
						],
						"type": "object"
					},
					"itens": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"id",
					"valueTotal",
					"description",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_SaleDetailResponse_": {
				"properties": {
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/SaleDetailResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"totalItems",
					"totalPages",
					"currentPage",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaleUpdateRequest": {
				"properties": {
					"description": {
						"type": "string",
						"example": "Nova descrição da venda"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ItemResponse": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Teclado Mecânico"
					},
					"description": {
						"type": "string",
						"example": "Teclado RGB switch blue"
					},
					"price": {
						"type": "number",
						"format": "double",
						"example": 250.5
					},
					"quantity": {
						"type": "number",
						"format": "double",
						"example": 10
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"price",
					"quantity",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemRequest": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Teclado Mecânico"
					},
					"description": {
						"type": "string",
						"example": "Teclado RGB switch blue"
					},
					"price": {
						"type": "number",
						"format": "double",
						"example": 250.5
					},
					"quantity": {
						"type": "number",
						"format": "double",
						"example": 10
					}
				},
				"required": [
					"name",
					"description",
					"price",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_ItemResponse_": {
				"properties": {
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"totalItems",
					"totalPages",
					"currentPage",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ItemRequest_": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Teclado Mecânico"
					},
					"description": {
						"type": "string",
						"example": "Teclado RGB switch blue"
					},
					"price": {
						"type": "number",
						"format": "double",
						"example": 250.5
					},
					"quantity": {
						"type": "number",
						"format": "double",
						"example": 10
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "API do sistema de vendas com geração automática via TSOA",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login efetuado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					},
					"400": {
						"description": "Credenciais inválidas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Realiza a autenticação do usuário e retorna um token JWT.",
				"tags": [
					"Autenticação"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/users": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Criado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Erro de validação",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usuários"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_UserResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Usuários"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseDTO"
								}
							}
						}
					},
					"404": {
						"description": "Não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usuários"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Erro de validação",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usuários"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usuários"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/sales": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Venda criada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro de negócio",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vendas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaleCreateRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_SaleDetailResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Vendas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/sales/admin/all": {
			"get": {
				"operationId": "GetAllForAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_SaleDetailResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Vendas"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/sales/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaleDetailResponse"
								}
							}
						}
					},
					"404": {
						"description": "Venda não encontrada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vendas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro de validação",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Venda não encontrada ou acesso negado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vendas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaleUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/items": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Criado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro de validação",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Itens"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ItemRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_ItemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Itens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/items/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemResponse"
								}
							}
						}
					},
					"404": {
						"description": "Não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Itens"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemResponse"
								}
							}
						}
					},
					"404": {
						"description": "Não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Itens"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ItemRequest_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Erro ao deletar",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Itens"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}